添加联系方式设置

+++ game-core/game-core-dao/src/main/java/com/fulu/game/core/dao/GradingPriceDao.java
@@ -21,7 +21,7 @@ public interface GradingPriceDao extends ICommonDao<GradingPrice,Integer>{
     List<GradingPrice> findByParameter(GradingPriceVO gradingPriceVO);
 
 
-    BigDecimal findRangePrice(@Param(value = "categoryId") Integer categoryId,
+    Double findRangePrice(@Param(value = "categoryId") Integer categoryId,
                               @Param(value = "startRank")  Integer startRank,
                               @Param(value = "endRank") Integer endRank);
